Split out scope analysis module
