+++ src/lib.rs
@@ -7,6 +7,7 @@ extern crate lazy_static;
 
 mod syntax;
 mod preprocess;
+mod scope;
 
 use walkdir::WalkDir;
 
@@ -14,10 +15,7 @@ use failure::Error;
 use failure::ResultExt;
 
 use std::collections::BTreeMap as Map;
-use std::collections::BTreeSet as Set;
 use std::path::{Path, PathBuf};
-
-use preprocess::Parsed;
 
 pub fn run(root_path: impl AsRef<Path>, emitter: &mut Emitter) -> Result<(), Error> {
     run_(root_path.as_ref(), emitter)
@@ -54,7 +52,7 @@ fn run_(root_path: &Path, emitter: &mut Emitter) -> Result<(), Error> {
         };
     }
 
-    analyze(&files, emitter).context("analyzing")?;
+    scope::analyze(&files, emitter).context("analyzing")?;
 
     Ok(())
 }
@@ -141,145 +139,4 @@ impl Emitter for VecEmitter {
         let to_emit = EmittedItem { kind, message, location, notes };
         self.emitted_items.push(to_emit)
     }
-}
-
-// ---- Scope analysis -------------------------------------------------------
-
-/// Functions in scope
-#[derive(Debug, Clone, Default)]
-struct Scope<'a> {
-    /// Functions defined in this scopeanalyze
-    defined: Set<&'a str>,
-    /// Defined by a file imported by `analyze
-    directly_imported: Set<&'a str>,
-    /// All the functions in scope
-    all: Set<&'a str>,
-    /// All the files imported (directlanalyzely)
-    files: Set<&'a Path>,
-}
-
-/// Type of function found in scope
-#[derive(Debug)]
-enum Found {
-    /// Found in Scope::defined or Scope::indirectly_imported
-    Direct,
-    /// Indirectly imported (through multiple layers of `.`)
-    Indirect,
-}
-
-impl<'a> Scope<'a> {
-    fn search(&self, name: &str) -> Option<Found> {
-        if self.all.contains(name) {
-            if self.defined.contains(name) || self.directly_imported.contains(name) {
-                Some(Found::Direct)
-            } else {
-                Some(Found::Indirect)
-            }
-        } else {
-            None
-        }
-    }
-}
-
-/// State of scope computation
-#[derive(Debug, Clone)]
-enum ScopeWip<'a> {
-    /// Done
-    Resolved(Scope<'a>),
-
-    /// The scope is being currently computed
-    /// (used to detect import loop)
-    Current,
-}
-
-fn analyze(files: &Map<PathBuf, Parsed>, emitter: &mut Emitter) -> Result<(), Error> {
-    lazy_static! {
-        static ref BUILTINS: Set<&'static str> = include_str!("builtins.txt")
-            .split_whitespace()
-            .chain(include_str!("extras.txt").split_whitespace())
-            .collect();
-    }
-
-    let mut scopes = Map::new();
-
-    for path in files.keys() {
-        let scope = get_scope(path, files, &mut scopes)?;
-
-        let parsed = &files[path];
-
-        let mut already_analyzed = Set::new();
-
-        for usage in &parsed.usages {
-            if BUILTINS.contains(usage.name.as_str()) {
-                continue;
-            }
-            if is_allowed(&usage.location.line, &usage.name) {
-                continue;
-            }
-            if already_analyzed.contains(usage.name.as_str()) {
-                continue;
-            }
-
-            already_analyzed.insert(usage.name.as_str());
-
-            match scope.search(&usage.name) {
-                None => emitter.emit(
-                    Message::Error,
-                    format!("Not in scope: {}", usage.name),
-                    usage.location.in_file(&parsed.original_path),
-                    None,
-                ),
-                Some(Found::Indirect) => emitter.emit(
-                    Message::Warning,
-                    format!("Indirectly imported: {}", usage.name),
-                    usage.location.in_file(&parsed.original_path),
-                    None,
-                ),
-                _ => (),
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn get_scope<'a>(
-    file: &'a Path,
-    files: &'a Map<PathBuf, Parsed>,
-    scopes: &mut Map<&'a Path, ScopeWip<'a>>,
-) -> Result<Scope<'a>, Error> {
-    match scopes.get(file) {
-        Some(ScopeWip::Current) => bail!("Recursive import of {}", file.display()),
-        Some(ScopeWip::Resolved(scope)) => return Ok(scope.clone()),
-        _ => (),
-    };
-    scopes.insert(file, ScopeWip::Current);
-
-    let parsed_file = files.get(file).ok_or_else(|| {
-        format_err!(
-            "List of elements in scope of file {} was requested, \
-             but not available due to previous import error",
-            file.display()
-        )
-    })?;
-
-    let mut scope = Scope::default();
-
-    for import in &parsed_file.imports {
-        let nested = get_scope(&import, files, scopes)?;
-        scope.directly_imported.extend(&nested.defined);
-        scope.all.extend(&nested.all);
-        scope.files.extend(&nested.files);
-    }
-
-    for definition in &parsed_file.definitions {
-        scope.defined.insert(&definition.name);
-        scope.all.insert(&definition.name);
-    }
-
-    scope.files.insert(file);
-
-    scopes.insert(file, ScopeWip::Resolved(scope.clone()));
-
-    Ok(scope)
 }
